requirements.txt
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
